% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_tidy-data.R
\name{plot_character_barplots}
\alias{plot_character_barplots}
\title{Plot bar plots for all character columns}
\usage{
plot_character_barplots(df, output_dir)
}
\arguments{
\item{df}{Data frame to plot}

\item{output_dir}{Output directory path (will be created if needed)}
}
\description{
Plot bar plots for all character columns
}
