% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{plot_qq}
\alias{plot_qq}
\title{Plot QQ plot of residuals}
\usage{
plot_qq(residuals, output_path = NULL)
}
\arguments{
\item{residuals}{Numeric vector of residuals}

\item{output_path}{Optional path to save plot}
}
\value{
ggplot object
}
\description{
Plot QQ plot of residuals
}
