% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{plot_residuals_histogram}
\alias{plot_residuals_histogram}
\title{Plot histogram of residuals}
\usage{
plot_residuals_histogram(residuals, output_path = NULL)
}
\arguments{
\item{residuals}{Numeric vector of residuals}

\item{output_path}{Optional path to save plot}
}
\value{
ggplot object
}
\description{
Plot histogram of residuals
}
