% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{range_stdize}
\alias{range_stdize}
\title{Range standardization (Steinley 2004, 2006)
zi = (xi - min(x)) / (max(x) - min(x))
Stores attributes for reverse transformation}
\usage{
range_stdize(x)
}
\arguments{
\item{x}{Numeric vector to standardize}
}
\value{
Range-standardized vector (0-1) with attributes for reversal
}
\description{
Range standardization (Steinley 2004, 2006)
zi = (xi - min(x)) / (max(x) - min(x))
Stores attributes for reverse transformation
}
