% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{stdize}
\alias{stdize}
\title{Standardize to 0-1 scale (centre, scale to SD, rescale)
Centres around mean, scales to SD, then rescales to 0:1
Stores attributes for reverse transformation}
\usage{
stdize(x)
}
\arguments{
\item{x}{Numeric vector to standardize}
}
\value{
Standardized vector with attributes for reversal
}
\description{
Standardize to 0-1 scale (centre, scale to SD, rescale)
Centres around mean, scales to SD, then rescales to 0:1
Stores attributes for reverse transformation
}
