% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{test_homogeneity}
\alias{test_homogeneity}
\title{Test homogeneity of variance
Runs Bartlett and Levene tests}
\usage{
test_homogeneity(formula, data)
}
\arguments{
\item{formula}{Formula: response ~ group}

\item{data}{Data frame}
}
\value{
List with test results
}
\description{
Test homogeneity of variance
Runs Bartlett and Levene tests
}
