% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{transform_arcsin}
\alias{transform_arcsin}
\title{Arcsin (angular) transformation for proportions}
\usage{
transform_arcsin(x, n = NULL)
}
\arguments{
\item{x}{Numeric vector of proportions (0-1)}

\item{n}{Sample size (for edge corrections)}
}
\value{
Transformed vector
}
\description{
Arcsin (angular) transformation for proportions
}
