% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{transform_power}
\alias{transform_power}
\title{Power transformation}
\usage{
transform_power(x, power = 2)
}
\arguments{
\item{x}{Numeric vector}

\item{power}{Power to raise to (default: 2)}
}
\value{
Transformed vector with attributes
}
\description{
Power transformation
}
