% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{transform_reciprocal}
\alias{transform_reciprocal}
\title{Reciprocal transformation}
\usage{
transform_reciprocal(x, add_constant = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{add_constant}{Value to add if zeros (default: 1)}
}
\value{
Transformed vector with attributes
}
\description{
Reciprocal transformation
}
