% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{transform_sqrt}
\alias{transform_sqrt}
\title{Square root transformation}
\usage{
transform_sqrt(x, add_constant = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{add_constant}{Value to add if zeros/negatives (default: 0)}
}
\value{
Transformed vector with attributes
}
\description{
Square root transformation
}
