% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{unrange_stdize}
\alias{unrange_stdize}
\title{Reverse range standardization}
\usage{
unrange_stdize(x)
}
\arguments{
\item{x}{Range-standardized vector (with attributes from range_stdize)}
}
\value{
Original scale values
}
\description{
Reverse range standardization
}
