% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{unstdize}
\alias{unstdize}
\title{Reverse stdize transformation}
\usage{
unstdize(x)
}
\arguments{
\item{x}{Standardized vector (with attributes from stdize)}
}
\value{
Original scale values
}
\description{
Reverse stdize transformation
}
