% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{untransform_arcsin}
\alias{untransform_arcsin}
\title{Reverse arcsin transformation}
\usage{
untransform_arcsin(x)
}
\arguments{
\item{x}{Arcsin-transformed vector}
}
\value{
Original proportions
}
\description{
Reverse arcsin transformation
}
