% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{untransform_log}
\alias{untransform_log}
\title{Reverse log transformation}
\usage{
untransform_log(x)
}
\arguments{
\item{x}{Log-transformed vector}
}
\value{
Original scale values
}
\description{
Reverse log transformation
}
