% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{untransform_power}
\alias{untransform_power}
\title{Reverse power transformation}
\usage{
untransform_power(x)
}
\arguments{
\item{x}{Power-transformed vector}
}
\value{
Original scale values
}
\description{
Reverse power transformation
}
