% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{untransform_reciprocal}
\alias{untransform_reciprocal}
\title{Reverse reciprocal transformation}
\usage{
untransform_reciprocal(x)
}
\arguments{
\item{x}{Reciprocal-transformed vector}
}
\value{
Original scale values
}
\description{
Reverse reciprocal transformation
}
