% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_transform.R
\name{untransform_sqrt}
\alias{untransform_sqrt}
\title{Reverse square root transformation}
\usage{
untransform_sqrt(x)
}
\arguments{
\item{x}{Sqrt-transformed vector}
}
\value{
Original scale values
}
\description{
Reverse square root transformation
}
